allow re-using logic for resolving k8s container images
